routes: make RouteState an abstract class
