+++ src/main/java/fr/sncf/osrd/infra/railscript/value/RSType.java
@@ -1,7 +1,7 @@
 package fr.sncf.osrd.infra.railscript.value;
 
 import fr.sncf.osrd.infra.InvalidInfraException;
-import fr.sncf.osrd.infra_state.RouteStatus;
+import fr.sncf.osrd.infra_state.routes.RouteStatus;
 
 public enum RSType {
     BOOLEAN(null),
