+++ src/main/java/fr/sncf/osrd/railjson/parser/RailScriptExprParser.java
@@ -10,8 +10,8 @@ import fr.sncf.osrd.infra.railscript.value.RSBool;
 import fr.sncf.osrd.infra.railscript.value.RSOptional;
 import fr.sncf.osrd.infra.railscript.value.RSType;
 import fr.sncf.osrd.infra.signaling.Aspect;
-import fr.sncf.osrd.infra_state.RouteState;
-import fr.sncf.osrd.infra_state.RouteStatus;
+import fr.sncf.osrd.infra_state.routes.RouteState;
+import fr.sncf.osrd.infra_state.routes.RouteStatus;
 import fr.sncf.osrd.infra_state.SignalState;
 import fr.sncf.osrd.railjson.schema.infra.RJSRoute;
 import fr.sncf.osrd.railjson.schema.infra.railscript.RJSRSExpr;
